backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg[binary]==3.1.18
python-multipart==0.0.6
chromadb==0.4.18
openai==1.3.5
google-generativeai==0.3.1
PyMuPDF==1.23.8
python-dotenv==1.0.0
httpx==0.25.2
